+++ dhlab/text/corpus.py
@@ -1,6 +1,9 @@
+from typing import Union
+
+import pandas as pd
 from pandas import DataFrame
 
-from dhlab.api.dhlab_api import document_corpus, get_metadata
+from dhlab.api.dhlab_api import document_corpus, get_metadata, evaluate_documents
 
 
 class Corpus:
@@ -18,8 +21,7 @@ class Corpus:
             ddk=None,
             subject=None,
             lang=None,
-            limit=None):
-
+            limit=10):
         self.corpus = document_corpus(
             doctype,
             author,
@@ -38,12 +40,47 @@ class Corpus:
 
         self.size = len(self.corpus)
 
-        return
+    def __repr__(self) -> str:
+        """
+        Return the string representation of the corpus datafrane
+        """
+        return self.corpus.__repr__()
+
+    def _repr_html_(self) -> Union[str, None]:
+        """
+        Return the HTML representation of the corpus datafrane
+        """
+        return self.corpus._repr_html_()
+
+    def add(self, corpus=None):
+        """Add a corpus to existing corpus"""
+        self.corpus = pd.concat([self.corpus, corpus.corpus]).drop_duplicates().reset_index(
+            drop=True)
+        self.size = len(self.corpus)
+
+    def extend_from_identifiers(self, identifiers=None):
+        other_corpus = get_metadata(urnlist(identifiers))
+        self.add(other_corpus)
+
+    def evaluate_words(self, wordbags=None):
+        df = evaluate_documents(wordbags=wordbags, urns=list(self.corpus.urn))
+        df.index = df.index.astype(int)
+        cols = df.columns
+        df = pd.concat([df, self.corpus[['dhlabid', 'urn']].set_index('dhlabid')], axis=1)
+        df = df.set_index('urn')
+        return df[cols].fillna(0)
 
 
-class CorpusFromIdentifiers(Corpus):
+class EmptyCorpus(Corpus):
+    def __init__(self):
+        self.corpus = pd.DataFrame()
+        self.size = 0
+
+
+class Corpus_from_identifiers(Corpus):
     def __init__(self, identifiers=None):
         self.corpus = get_metadata(urnlist(identifiers))
+        self.size = len(self.corpus)
 
 
 def urnlist(corpus):